feat: S1.5 MessageId and Message — fields on SolidSyslogMessage (#47)
* feat: S1.5 BDD scaffold — message fields feature, steps, test list

BDD feature file for messageId and message body with steps for
--msgid and --message CLI args. IGNORE_TEST with planned TDD list.


* refactor: message as test group member, remove Log overloads

SolidSyslogMessage is now a member initialised in setup. Tests
set individual fields before calling Log(). Simplifies the
three-line arrange/act/assert pattern. FormatMsgId extracted
with StringIsValid helper.


* feat: S1.5 messageId — zero, one, boundary tests with truncation

FormatMsgId with StringIsValid, FormatBoundedString for 32-char
RFC limit. Empty/NULL produces NILVALUE. CHECK_MSGID macro.


* feat: S1.5 message body — NULL/empty omits MSG, FormatMsg/FormatStructuredData extracted


* feat: S1.5 message body — boundary tests, FormatMsg with remaining size

MSG truncated at SOLIDSYSLOG_MAX_MESSAGE_SIZE (now public in
SolidSyslog.h). FormatMsg receives remaining buffer space.
Tests: spaces preserved, fills buffer, truncation, huge message
does not corrupt.


* refactor: change Format functions and StringFunction to return size_t

Eliminates sign-conversion casts throughout. len variables are
size_t consistently. SolidSyslogStringFunction typedef, POSIX
helpers, StringFake, and example all updated. FormatString removed
(unused). FormatBoundedString reserves space for null terminator
and checks bounds before array access.


* feat: S1.5 integration — msgid/message CLI flags, BDD updated, cleanup

Example program accepts --msgid and --message CLI flags. Walking
skeleton BDD updated for NULL defaults. Superseded IGNORE_TESTs
and completed test list removed. size_t overflow fix in
FormatBoundedString. Unused FormatString removed.


* fix: resolve Behave AmbiguousStep for combined msgid+body scenario

Rename step to 'sends a complete message with...' to avoid
pattern collision with 'sends a message with message ID...'.


* fix: remove MSGID assertion from walking skeleton

syslog-ng omits MSGID from output when value is NILVALUE,
causing KeyError in the step parser. MSGID is tested in the
message_fields feature where it has a real value.


---------

+++ Interface/SolidSyslogConfig.h
@@ -9,7 +9,7 @@ EXTERN_C_BEGIN
 
     struct SolidSyslogSender;
 
-    typedef int (*SolidSyslogStringFunction)(char* buffer, size_t size);
+    typedef size_t (*SolidSyslogStringFunction)(char* buffer, size_t size);
 
     struct SolidSyslogConfig
     {
