+++ Tests/StringFake.h
@@ -5,13 +5,13 @@
 
 EXTERN_C_BEGIN
 
-    void StringFake_Reset(void);
-    void StringFake_SetHostname(const char* hostname);
-    int  StringFake_GetHostname(char* buffer, size_t size);
-    void StringFake_SetAppName(const char* appName);
-    int  StringFake_GetAppName(char* buffer, size_t size);
-    void StringFake_SetProcId(const char* procId);
-    int  StringFake_GetProcId(char* buffer, size_t size);
+    void   StringFake_Reset(void);
+    void   StringFake_SetHostname(const char* hostname);
+    size_t StringFake_GetHostname(char* buffer, size_t size);
+    void   StringFake_SetAppName(const char* appName);
+    size_t StringFake_GetAppName(char* buffer, size_t size);
+    void   StringFake_SetProcId(const char* procId);
+    size_t StringFake_GetProcId(char* buffer, size_t size);
 
 EXTERN_C_END
