+++ Tests/StringFakeTest.cpp
@@ -14,8 +14,8 @@ TEST_GROUP(StringFake)
 
 TEST(StringFake, ReturnsEmptyStringAfterReset)
 {
-    char buffer[32];
-    int  len = StringFake_GetHostname(buffer, sizeof(buffer));
+    char   buffer[32];
+    size_t len = StringFake_GetHostname(buffer, sizeof(buffer));
     STRCMP_EQUAL("", buffer);
     LONGS_EQUAL(0, len);
 }
@@ -23,8 +23,8 @@ TEST(StringFake, ReturnsEmptyStringAfterReset)
 TEST(StringFake, ReturnsConfiguredHostname)
 {
     StringFake_SetHostname("MyHost");
-    char buffer[32];
-    int  len = StringFake_GetHostname(buffer, sizeof(buffer));
+    char   buffer[32];
+    size_t len = StringFake_GetHostname(buffer, sizeof(buffer));
     STRCMP_EQUAL("MyHost", buffer);
     LONGS_EQUAL(6, len);
 }
@@ -32,8 +32,8 @@ TEST(StringFake, ReturnsConfiguredHostname)
 TEST(StringFake, TruncatesWhenBufferTooSmall)
 {
     StringFake_SetHostname("LongHostname");
-    char buffer[5];
-    int  len = StringFake_GetHostname(buffer, sizeof(buffer));
+    char   buffer[5];
+    size_t len = StringFake_GetHostname(buffer, sizeof(buffer));
     STRCMP_EQUAL("Long", buffer);
     LONGS_EQUAL(4, len);
 }
